django_ecommerce/requirements.txt
Django==3.2.7
django-admin-honeypot==1.1.0
django-admin-thumbnails==0.2.6
django-session-timeout==0.1.0
Pillow==8.2.0